+++ libs/components/navbar/projects/navbar/src/public-api.ts
@@ -1,0 +1,6 @@
+export * from './modules/navbar/navbar.module';
+
+// Components and directives must be exported to support Angular's "partial" Ivy compiler.
+// Obscure names are used to indicate types are not part of the public API.
+export { SkyNavbarComponent as λ1 } from './modules/navbar/navbar.component';
+export { SkyNavbarItemComponent as λ2 } from './modules/navbar/navbar-item.component';