+++ libs/components/navbar/src/app/visual/navbar/navbar-visual.component.ts
@@ -1,9 +1,0 @@
-import {
-  Component
-} from '@angular/core';
-
-@Component({
-  selector: 'navbar-visual',
-  templateUrl: './navbar-visual.component.html'
-})
-export class NavbarVisualComponent { }