+++ libs/components/navbar/CHANGELOG.md
@@ -1,3 +1,19 @@
+# 5.0.0-beta.3 (2021-09-10)
+
+- Migrated to Angular CLI. [#20](https://github.com/blackbaud/skyux-navbar/pull/20)
+
+# 5.0.0-beta.2 (2021-07-15)
+
+- Added support for `5.0.0-beta.*` versions of SKY UX component libraries. [#19](https://github.com/blackbaud/skyux-navbar/pull/19)
+
+# 5.0.0-beta.1 (2021-06-25)
+
+- Added support for `@angular/core@^12` and `@skyux/i18n@^5.0.0-beta.1`. [#18](https://github.com/blackbaud/skyux-navbar/pull/18)
+
+# 5.0.0-beta.0 (2021-06-24)
+
+- Initial 5.0.0-beta release.
+
 # 4.0.0 (2020-05-12)
 
 ### New features